separates words_service into new file
resolves #70
